Organize `generator` files a bit
