+++ test/fixtures/generator/bwx_spec.json
@@ -445,46 +445,47 @@
             "info_plist": null,
             "inputs": {
                 "srcs": [
-                    "tools/generator/src/BuildMode.swift",
-                    "tools/generator/src/BuildSetting.swift",
                     "tools/generator/src/BuildSettingConditional.swift",
-                    "tools/generator/src/ConsolidatedTarget+Extensions.swift",
-                    "tools/generator/src/DTO.swift",
-                    "tools/generator/src/Environment.swift",
+                    "tools/generator/src/DTO/BuildMode.swift",
+                    "tools/generator/src/DTO/BuildSetting.swift",
+                    "tools/generator/src/DTO/FilePath.swift",
+                    "tools/generator/src/DTO/Inputs.swift",
+                    "tools/generator/src/DTO/LinkerInputs.swift",
+                    "tools/generator/src/DTO/Outputs.swift",
+                    "tools/generator/src/DTO/Platform.swift",
+                    "tools/generator/src/DTO/Product.swift",
+                    "tools/generator/src/DTO/Project.swift",
+                    "tools/generator/src/DTO/SearchPaths.swift",
+                    "tools/generator/src/DTO/Target.swift",
+                    "tools/generator/src/DTO/TargetID.swift",
+                    "tools/generator/src/DTO/XCCurrentVersion.swift",
                     "tools/generator/src/Errors.swift",
-                    "tools/generator/src/FilePath.swift",
+                    "tools/generator/src/Extensions/PBXGroup+Extensions.swift",
+                    "tools/generator/src/Extensions/PBXProductType+Extensions.swift",
+                    "tools/generator/src/Extensions/PBXTarget+Extensions.swift",
+                    "tools/generator/src/Extensions/Path+Extensions.swift",
+                    "tools/generator/src/Extensions/Sorting.swift",
+                    "tools/generator/src/Extensions/XCSchemeEnvrionmentVariables+Extensions.swift",
                     "tools/generator/src/FilePathResolver.swift",
-                    "tools/generator/src/Generator+AddBazelDependenciesTarget.swift",
-                    "tools/generator/src/Generator+AddTargets.swift",
-                    "tools/generator/src/Generator+ConsolidateTargets.swift",
-                    "tools/generator/src/Generator+CreateFilesAndGroups.swift",
-                    "tools/generator/src/Generator+CreateProducts.swift",
-                    "tools/generator/src/Generator+CreateProject.swift",
-                    "tools/generator/src/Generator+CreateXCSchemes.swift",
-                    "tools/generator/src/Generator+CreateXCSharedData.swift",
-                    "tools/generator/src/Generator+CreateXcodeProj.swift",
-                    "tools/generator/src/Generator+DisambiguateTargets.swift",
-                    "tools/generator/src/Generator+Main.swift",
-                    "tools/generator/src/Generator+PopulateMainGroup.swift",
-                    "tools/generator/src/Generator+ProcessTargetMerges.swift",
-                    "tools/generator/src/Generator+SetTargetConfigurations.swift",
-                    "tools/generator/src/Generator+SetTargetDependencies.swift",
-                    "tools/generator/src/Generator+WriteXcodeProj.swift",
-                    "tools/generator/src/Generator.swift",
-                    "tools/generator/src/Inputs.swift",
-                    "tools/generator/src/LinkerInputs.swift",
-                    "tools/generator/src/Logger.swift",
-                    "tools/generator/src/Outputs.swift",
-                    "tools/generator/src/PBXGroup+Extensions.swift",
-                    "tools/generator/src/PBXProductType+Extensions.swift",
-                    "tools/generator/src/PBXTarget+Extensions.swift",
-                    "tools/generator/src/PBXTargets+Extensions.swift",
-                    "tools/generator/src/Path+Extensions.swift",
-                    "tools/generator/src/SearchPaths.swift",
-                    "tools/generator/src/Sorting.swift",
-                    "tools/generator/src/TargetID.swift",
-                    "tools/generator/src/XCCurrentVersion.swift",
-                    "tools/generator/src/XCSchemeEnvrionmentVariables+Extensions.swift"
+                    "tools/generator/src/Generator/AddBazelDependenciesTarget.swift",
+                    "tools/generator/src/Generator/AddTargets.swift",
+                    "tools/generator/src/Generator/ConsolidateTargets.swift",
+                    "tools/generator/src/Generator/CreateFilesAndGroups.swift",
+                    "tools/generator/src/Generator/CreateProducts.swift",
+                    "tools/generator/src/Generator/CreateProject.swift",
+                    "tools/generator/src/Generator/CreateXCSchemes.swift",
+                    "tools/generator/src/Generator/CreateXCSharedData.swift",
+                    "tools/generator/src/Generator/CreateXcodeProj.swift",
+                    "tools/generator/src/Generator/DisambiguateTargets.swift",
+                    "tools/generator/src/Generator/Environment.swift",
+                    "tools/generator/src/Generator/Generator.swift",
+                    "tools/generator/src/Generator/Main.swift",
+                    "tools/generator/src/Generator/PopulateMainGroup.swift",
+                    "tools/generator/src/Generator/ProcessTargetMerges.swift",
+                    "tools/generator/src/Generator/SetTargetConfigurations.swift",
+                    "tools/generator/src/Generator/SetTargetDependencies.swift",
+                    "tools/generator/src/Generator/WriteXcodeProj.swift",
+                    "tools/generator/src/Logger.swift"
                 ]
             },
             "is_swift": true,
