--- tools/generator/src/ConsolidatedTarget+Extensions.swift
@@ -1,13 +1,0 @@
-extension ConsolidatedTarget {
-    /// The normalized name is used during target disambiguation. It allows the
-    /// logic to differentiate targets where the names only differ by case.
-    var normalizedName: String {
-        return name.lowercased()
-    }
-
-    /// The normalized label is used during target disambiguation. It allows the
-    /// logic to differentiate targets where the names only differ by case.
-    var normalizedLabel: String {
-        return label.lowercased()
-    }
-}
