+++ tools/generator/src/DTO/BuildSetting.swift
