+++ tools/generator/src/DTO/Inputs.swift
