+++ tools/generator/src/DTO/LinkerInputs.swift
