+++ tools/generator/src/DTO/Outputs.swift
