+++ tools/generator/src/DTO/Platform.swift
@@ -1,46 +1,3 @@
-import PathKit
-import XcodeProj
-
-struct Project: Equatable, Decodable {
-    let name: String
-    let bazelWorkspaceName: String
-    let label: String
-    let configuration: String
-    let buildSettings: [String: BuildSetting]
-    var targets: [TargetID: Target]
-    let targetMerges: [TargetID: Set<TargetID>]
-    let invalidTargetMerges: [TargetID: Set<TargetID>]
-    let extraFiles: Set<FilePath>
-}
-
-struct Target: Equatable, Decodable {
-    let name: String
-    let label: String
-    let configuration: String
-    var packageBinDir: Path
-    let platform: Platform
-    let product: Product
-    var isSwift: Bool
-    let testHost: TargetID?
-    var buildSettings: [String: BuildSetting]
-    var searchPaths: SearchPaths
-    var modulemaps: [FilePath]
-    var swiftmodules: [FilePath]
-    let resourceBundles: Set<FilePath>
-    var inputs: Inputs
-    var linkerInputs: LinkerInputs
-    var infoPlist: FilePath?
-    var entitlements: FilePath?
-    var dependencies: Set<TargetID>
-    var outputs: Outputs
-}
-
-struct Product: Equatable, Decodable {
-    let type: PBXProductType
-    let name: String
-    let path: FilePath
-}
-
 struct Platform: Equatable, Hashable, Decodable {
     enum OS: String, Decodable {
         case macOS = "macos"
@@ -55,6 +12,8 @@ struct Platform: Equatable, Hashable, Decodable {
     let minimumOsVersion: String
     let environment: String?
 }
+
+// MARK: - Comparable
 
 extension Platform: Comparable {
     static func < (lhs: Platform, rhs: Platform) -> Bool {
