+++ tools/generator/src/DTO/Product.swift
@@ -1,0 +1,7 @@
+import XcodeProj
+
+struct Product: Equatable, Decodable {
+    let type: PBXProductType
+    let name: String
+    let path: FilePath
+}
