+++ tools/generator/src/DTO/SearchPaths.swift
