+++ tools/generator/src/DTO/Target.swift
@@ -1,0 +1,23 @@
+import PathKit
+
+struct Target: Equatable, Decodable {
+    let name: String
+    let label: String
+    let configuration: String
+    var packageBinDir: Path
+    let platform: Platform
+    let product: Product
+    var isSwift: Bool
+    let testHost: TargetID?
+    var buildSettings: [String: BuildSetting]
+    var searchPaths: SearchPaths
+    var modulemaps: [FilePath]
+    var swiftmodules: [FilePath]
+    let resourceBundles: Set<FilePath>
+    var inputs: Inputs
+    var linkerInputs: LinkerInputs
+    var infoPlist: FilePath?
+    var entitlements: FilePath?
+    var dependencies: Set<TargetID>
+    var outputs: Outputs
+}
