+++ tools/generator/src/DTO/XCCurrentVersion.swift
