+++ tools/generator/src/Extensions/PBXGroup+Extensions.swift
