+++ tools/generator/src/Extensions/PBXProductType+Extensions.swift
