+++ tools/generator/src/Extensions/Path+Extensions.swift
