+++ tools/generator/src/Extensions/XCSchemeEnvrionmentVariables+Extensions.swift
