+++ tools/generator/src/Generator/AddBazelDependenciesTarget.swift
