+++ tools/generator/src/Generator/AddTargets.swift
