+++ tools/generator/src/Generator/ConsolidateTargets.swift
