+++ tools/generator/src/Generator/CreateFilesAndGroups.swift
