+++ tools/generator/src/Generator/CreateXCSharedData.swift
