+++ tools/generator/src/Generator/DisambiguateTargets.swift
@@ -472,6 +472,20 @@ struct EnvironmentSystemComponents {
     }
 }
 
+extension ConsolidatedTarget {
+    /// The normalized name is used during target disambiguation. It allows the
+    /// logic to differentiate targets where the names only differ by case.
+    var normalizedName: String {
+        return name.lowercased()
+    }
+
+    /// The normalized label is used during target disambiguation. It allows the
+    /// logic to differentiate targets where the names only differ by case.
+    var normalizedLabel: String {
+        return label.lowercased()
+    }
+}
+
 private extension Target {
     /// A key that corresponds to the most-distinguished string that
     /// `ProductTypeComponents.distinguisher()` can return for this
