+++ tools/generator/src/Generator/PopulateMainGroup.swift
