+++ tools/generator/src/Generator/ProcessTargetMerges.swift
