+++ tools/generator/src/Generator/SetTargetConfigurations.swift
