+++ tools/generator/src/Generator/SetTargetDependencies.swift
