--- tools/generator/src/PBXTargets+Extensions.swift
@@ -1,7 +1,0 @@
-import XcodeProj
-
-extension Dictionary where Value: PBXTarget {
-    func nativeTarget(_ targetID: Self.Key) -> PBXNativeTarget? {
-        return self[targetID] as? PBXNativeTarget
-    }
-}
